@@ -1,7 +1,7 @@
 {
   "author": "Audacia",
   "name": "@audacia/angular",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -13,29 +13,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.13",
-    "@angular/cli": "8.3.17",
-    "@angular/common": "8.2.13",
-    "@angular/compiler": "8.2.13",
-    "@angular/core": "8.2.13",
-    "@angular/forms": "8.2.13",
-    "@angular/platform-browser": "8.2.13",
-    "@angular/platform-browser-dynamic": "8.2.13",
-    "@angular/router": "8.2.13",
+    "@angular/animations": "9.1.13",
+    "@angular/cli": "9.1.15",
+    "@angular/common": "9.1.13",
+    "@angular/compiler": "9.1.13",
+    "@angular/core": "9.1.13",
+    "@angular/forms": "9.1.13",
+    "@angular/platform-browser": "9.1.13",
+    "@angular/platform-browser-dynamic": "9.1.13",
+    "@angular/router": "9.1.13",
     "rxjs": "6.5.3",
-    "tslib": "1.10.0",
-    "zone.js": "0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.803.17",
-    "@angular-devkit/build-ng-packagr": "0.803.17",
-    "@angular-devkit/schematics": "^16.2.3",
-    "@angular/compiler-cli": "8.2.13",
-    "@angular/language-service": "8.2.13",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular-devkit/build-ng-packagr": "~0.901.15",
+    "@angular-devkit/schematics": "16.2.3",
+    "@angular/compiler-cli": "9.1.13",
+    "@angular/language-service": "9.1.13",
     "@types/jasmine": "3.3.16",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "8.9.5",
-    "codelyzer": "5.1.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "4.1.0",
@@ -43,11 +43,10 @@
     "karma-coverage-istanbul-reporter": "2.0.6",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.2",
-    "ng-packagr": "5.7.1",
+    "ng-packagr": "^9.0.0",
     "protractor": "5.4.0",
     "ts-node": "7.0.1",
-    "tsickle": "0.36.0",
     "tslint": "5.15.0",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3"
   }
 }
